Added random loot

+++ LS/header/Game/Level/ObjectFactory.h
@@ -76,7 +76,7 @@ private:
 	glm::vec3 calcRot(glm::ivec2 square);
 
 public:
-	void createRandomLoot(std::vector<lootData> loot, float totalValue);
+	void createRandomLoot(std::vector<lootData> loot, int totalValue);
 	ObjectFactory(EventManager *events, const std::string &resourcePath = "", std::string modelPath = "");
 	~ObjectFactory();
